+++ pyphi/combinatorics.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # combinatorics.py
 """Combinatorial utilities."""