+++ pyphi/deferred/__init__.py
@@ -1,2 +1,3 @@
+# -*- coding: utf-8 -*-
 # deferred/__init__.py
 """Provides deferred imports to minimize core runtime dependencies."""