+++ pyphi/deferred/ray.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # deferred/ray.py
 """Provide the optional Ray dependency."""
 
@@ -23,7 +24,7 @@ class NoRay:
         return False
 
     def __repr__(self):
-        return "<NoRay(): " + self.__doc__ + ">"
+        return '<NoRay(): ' + self.__doc__ + '>'
 
 
 try: