+++ pyphi/network_generator/unit_functions.py
@@ -56,14 +56,14 @@ def boolean_function(element, weights, state, on_inputs=(), **kwargs):
         bool: The output of the element.
     """
     if np.any((weights != 1) & (weights != 0)):
-        raise NotImplementedError("weights must be 0 or 1")
+        raise NotImplementedError('weights must be 0 or 1')
     if len(set(map(len, on_inputs))) != 1:
-        raise ValueError("on_inputs must all be the same length")
+        raise ValueError('on_inputs must all be the same length')
 
     inputs = tuple(utils.weighted_inputs(element, weights, state))
 
     if len(inputs) != len(next(iter(on_inputs), len(inputs))):
-        raise ValueError("nonzero input weights and on_input lengths must match")
+        raise ValueError('nonzero input weights and on_input lengths must match')
 
     return inputs in on_inputs