+++ pyphi/visualize/connectivity.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # visualize/connectivity.py
 """Visualize system connectivity information."""
 
@@ -5,10 +6,10 @@ import networkx as nx
 
 NODE_COLORS = {
     # (in subsystem, state)
-    (False, 0): "lightgrey",
-    (False, 1): "darkgrey",
-    (True, 0): "lightblue",
-    (True, 1): "darkblue",
+    (False, 0): 'lightgrey',
+    (False, 1): 'darkgrey',
+    (True, 0): 'lightblue',
+    (True, 1): 'darkblue',
 }
 
 
@@ -18,7 +19,7 @@ def plot_graph(g, **kwargs):
             with_labels=True,
             arrowsize=20,
             node_size=600,
-            font_color="white",
+            font_color='white',
         ),
         **kwargs,
     }
@@ -33,8 +34,8 @@ def plot_subsystem(subsystem, **kwargs):
     nx.relabel_nodes(
         g, dict(zip(range(subsystem.network.size), subsystem.node_labels)), copy=False
     )
-    if "node_color" not in kwargs:
-        kwargs["node_color"] = [
+    if 'node_color' not in kwargs:
+        kwargs['node_color'] = [
             NODE_COLORS[(i in subsystem.node_indices, subsystem.state[i])]
             for i in range(subsystem.network.size)
         ]