+++ test/test_registry.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import pytest
 
 from pyphi.registry import Registry
@@ -6,16 +7,16 @@ from pyphi.registry import Registry
 def test_registry():
     registry = Registry()
 
-    assert "DIFF" not in registry
+    assert 'DIFF' not in registry
     assert len(registry) == 0
 
-    @registry.register("DIFF")
+    @registry.register('DIFF')
     def difference(a, b):
         return a - b
 
-    assert "DIFF" in registry
+    assert 'DIFF' in registry
     assert len(registry) == 1
-    assert registry["DIFF"] == difference
+    assert registry['DIFF'] == difference
 
     with pytest.raises(KeyError):
-        registry["HEIGHT"]
+        registry['HEIGHT']