+++ test/test_subsystem.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import numpy as np
 import pytest
 
@@ -145,11 +146,11 @@ def test_cut_node_labels(s):
 
 
 def test_specify_elements_with_labels(standard):
-    network = Network(standard.tpm.tpm, node_labels=("A", "B", "C"))
-    subsystem = Subsystem(network, (0, 0, 0), ("B", "C"))
+    network = Network(standard.tpm.tpm, node_labels=('A', 'B', 'C'))
+    subsystem = Subsystem(network, (0, 0, 0), ('B', 'C'))
     assert subsystem.node_indices == (1, 2)
-    assert tuple(node.label for node in subsystem.nodes) == ("B", "C")
-    assert str(subsystem) == "Subsystem(B, C)"
+    assert tuple(node.label for node in subsystem.nodes) == ('B', 'C')
+    assert str(subsystem) == 'Subsystem(B, C)'
 
 
 def test_null_concept(s):